+++ CHANGES.rst
@@ -4,6 +4,9 @@ Release 9.1.1 (in development)
 Bugs fixed
 ----------
 
+* #2914: Add ``sphinx.util.nodes.make_index()`` for extensions to create
+  index entries and their cross-reference targets.
+
 * #14465: LaTeX: PDF build crash since LaTeX June 2026 release if tables are
   styled with ``'colorrows'`` (which is the default).
   Patch by Jean-François B.